+++ src/widgets/PendingApprovalWidget.php
@@ -30,7 +30,7 @@ class PendingApprovalWidget extends Widget
 
     public static function isSelectable(): bool
     {
-        return Craft::$app->getUser()->checkPermission('vouch-viewReviews');
+        return Craft::$app->getUser()->checkPermission('vouch-viewWidgets');
     }
 
     public function getTitle(): ?string