+++ src/widgets/TopReviewedElementsWidget.php
@@ -42,7 +42,7 @@ class TopReviewedElementsWidget extends Widget
 
     public static function isSelectable(): bool
     {
-        return Craft::$app->getUser()->checkPermission('vouch-viewReviews');
+        return Craft::$app->getUser()->checkPermission('vouch-viewWidgets');
     }
 
     public function getTitle(): ?string